requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
openai>=1.12.0
google-generativeai>=0.3.2
deepgram-sdk>=3.0.0
pandas>=2.2.0
scikit-learn>=1.4.0
python-multipart>=0.0.9
python-docx>=1.1.0
Pillow>=10.0.0
